requirements.txt
graphene==2.0
django
psycopg2
pytest
pytest-cov
pytest-django
django-debug-toolbar
